+++ prototyp.h
@@ -1,37 +1,37 @@
 
 /* asm.c */
 extern int	GetRec_asm(InRecord *rec);
-extern int	PutHead_asm(FILE *file, ulong addr, ulong hi);
-extern int	PutRec_asm(FILE *file, uchar *data, int recsize, ulong recstart);
+extern int	PutHead_asm(FILE *file, uint32_t addr, uint32_t hi);
+extern int	PutRec_asm(FILE *file, uint8_t *data, int recsize, uint32_t recstart);
 
 /* dio.c */
 extern int	GetRec_dio(InRecord *rec);
-extern int	PutRec_dio(FILE *file, uchar *data, int recsize, ulong recstart);
-extern int	PutHead_dio(FILE *file, ulong addr, ulong hi);
+extern int	PutRec_dio(FILE *file, uint8_t *data, int recsize, uint32_t recstart);
+extern int	PutHead_dio(FILE *file, uint32_t addr, uint32_t hi);
 extern int	PutFoot_dio(FILE *file);
 
 /* cpe.c */
 extern int	GetRec_cpe(InRecord *rec);
-extern int	PutRec_cpe(FILE *file, uchar *data, int recsize, ulong recstart);
-extern int	PutHead_cpe(FILE *file, ulong addr, ulong hi);
+extern int	PutRec_cpe(FILE *file, uint8_t *data, int recsize, uint32_t recstart);
+extern int	PutHead_cpe(FILE *file, uint32_t addr, uint32_t hi);
 extern int	PutFoot_cpe(FILE *file);
-extern int	PutXfer_cpe(FILE *file, ulong addr);
+extern int	PutXfer_cpe(FILE *file, uint32_t addr);
 
 /* dld.c */
 extern int	GetRec_dld(InRecord *rec);
-extern int	PutRec_dld(FILE *file, uchar *data, int recsize, ulong recstart);
+extern int	PutRec_dld(FILE *file, uint8_t *data, int recsize, uint32_t recstart);
 extern int	PutFoot_dld(FILE *file);
 
 /* coff.c */
-extern int	GetRec_coff(GPF *gpf, InRecord *rec, LogicalAddr lo, LogicalAddr hi, long reloaction );
+extern int	GetRec_coff(GPF *gpf, InRecord *rec, LogicalAddr lo, LogicalAddr hi, int32_t reloaction );
 
 #if INCLUDE_ELF
 /* elf.c */
-extern int	GetRec_elf(GPF *gpf, InRecord *rec, LogicalAddr lo, LogicalAddr hi, long relocation );
+extern int	GetRec_elf(GPF *gpf, InRecord *rec, LogicalAddr lo, LogicalAddr hi, int32_t relocation );
 #endif
 
 /* getfile.c */
-extern void	save_data(InRecord *rec, LogicalAddr userLo, LogicalAddr userHigh, long relocation, GPF *gpf);
+extern void	save_data(InRecord *rec, LogicalAddr userLo, LogicalAddr userHigh, int32_t relocation, GPF *gpf);
 extern int	getfile(char *fname, GPF *gpf);
 
 /* getform.c */
@@ -39,48 +39,48 @@ extern void	fileExtension(char *dest, char *filename, int maxlen);
 extern FileFormat	getFormat(char *filename);
 
 /* hexutl.c */
-extern int	strtobytes(uchar *str, int nbytes);
-extern int	strtohex(uchar *str, int nchars);
-extern void	hextobytes(uchar *hexstr, int nbytes);
-extern LogicalAddr	bytestoaddr(uchar *bytestr, int nbytes);
-extern LogicalAddr	hextoaddr(uchar *hexstr, int nnybs);
+extern int	strtobytes(uint8_t *str, int nbytes);
+extern int	strtohex(uint8_t *str, int nchars);
+extern void	hextobytes(uint8_t *hexstr, int nbytes);
+extern LogicalAddr	bytestoaddr(uint8_t *bytestr, int nbytes);
+extern LogicalAddr	hextoaddr(uint8_t *hexstr, int nnybs);
 
 /* image.c */
 extern void	imageInit(Image *image, int pagesize);
 extern void	imageFree(Image *image);
-extern void	imageWrite(Image *image, LogicalAddr addr, size_t nbytes, uchar *data);
-extern void	imageRead(Image *image, LogicalAddr addr, size_t nbytes, uchar *data);
-extern void	symbolWrite(Image *image, uchar *sym, int len);
+extern void	imageWrite(Image *image, LogicalAddr addr, size_t nbytes, uint8_t *data);
+extern void	imageRead(Image *image, LogicalAddr addr, size_t nbytes, uint8_t *data);
+extern void	symbolWrite(Image *image, uint8_t *sym, int len);
 extern int	init_reader(GPF *gpf);
-extern int	readImage(GPF *gpf, uchar *bufferPtr, ulong bufferSpace, ulong low_address, 
-				ulong high_address, ulong *new_low, int *bytesRead);
+extern int	readImage(GPF *gpf, uint8_t *bufferPtr, uint32_t bufferSpace, uint32_t low_address, 
+				uint32_t high_address, uint32_t *new_low, int *bytesRead);
 extern void	imageDump(Image *image, FILE *dump);
 extern int	imageCheck(Image *image, FILE *log);
 
 /* img.c */
 extern int	GetRec_img(InRecord *rec);
-extern int	PutRec_img(FILE *file, uchar *data, int recsize, ulong recstart);
+extern int	PutRec_img(FILE *file, uint8_t *data, int recsize, uint32_t recstart);
 
 /* intel.c */
 extern int	GetRec_intel(InRecord *rec);
 extern int	PutFoot_intel(FILE *file);
-extern int	PutRec_intel(FILE *file, uchar *data, int recsize, ulong recstart);
+extern int	PutRec_intel(FILE *file, uint8_t *data, int recsize, uint32_t recstart);
 
 /* lda.c */
 extern int	GetRec_lda(InRecord *record);
-extern int	PutRec_lda(FILE *file, uchar *data, int recsize, ulong recstart);
+extern int	PutRec_lda(FILE *file, uint8_t *data, int recsize, uint32_t recstart);
 extern int	PutFoot_lda(FILE *file);
 
 /* mac.c */
 extern int	macGetRec(InRecord *rec);
-extern int	PutHead_mac(FILE *file, ulong addr, ulong hi);
-extern int	PutRec_mac(FILE *file, uchar *data, int recsize, ulong recstart);
+extern int	PutHead_mac(FILE *file, uint32_t addr, uint32_t hi);
+extern int	PutRec_mac(FILE *file, uint8_t *data, int recsize, uint32_t recstart);
 
 /* mot.c */
-extern char	*to_hex(uchar value, char *str);
+extern char	*to_hex(uint8_t value, char *str);
 extern int	GetRec_mot(InRecord *rec);
 extern int	PutFoot_mot(FILE *file);
-extern int	PutRec_mot(FILE *file, uchar *data, int recsize, ulong recstart);
+extern int	PutRec_mot(FILE *file, uint8_t *data, int recsize, uint32_t recstart);
 
 /* parser.c */
 extern char	*sig(char *text);
@@ -101,14 +101,14 @@ extern void	purge_qa2(FILE *fin, FILE *fout);
 
 /* rom.c */
 extern int	GetRec_rom(InRecord *rec);
-extern int	PutRec_rom(FILE *file, uchar *data, int recsize, ulong recstart);
-extern int	PutHead_rom(FILE *file, ulong addr, ulong hi);
+extern int	PutRec_rom(FILE *file, uint8_t *data, int recsize, uint32_t recstart);
+extern int	PutHead_rom(FILE *file, uint32_t addr, uint32_t hi);
 
 /* tekhex.c */
 extern int	GetRec_tekhex(InRecord *rec);
 extern int	PutFoot_tekhex(FILE *file);
-extern int	PutSym_tekhex(FILE *file, uchar *data, int recsize);
-extern int	PutRec_tekhex(FILE *file, uchar *data, int recsize, ulong recstart);
+extern int	PutSym_tekhex(FILE *file, uint8_t *data, int recsize);
+extern int	PutRec_tekhex(FILE *file, uint8_t *data, int recsize, uint32_t recstart);
 
 /* varargs.c */
 extern void	err_exit(char *fmt, ...) __attribute__ ((__format__ (__printf__, 1, 0)));
@@ -129,5 +129,5 @@ extern int	lookup_token(char *token, ...);
 
 /* vlda.c */
 extern int	GetRec_vlda(InRecord *record);
-extern int	PutSym_vlda(FILE *file, uchar *data, int recsize);
-extern int	PutRec_vlda(FILE *file, uchar *data, int recsize, ulong recstart);
+extern int	PutSym_vlda(FILE *file, uint8_t *data, int recsize);
+extern int	PutRec_vlda(FILE *file, uint8_t *data, int recsize, uint32_t recstart);
